fix(clientes): corrige catalogo de equipamentos e ajustes no cadastro

+++ backend/src/main/java/com/energ_ia/api/dto/cliente/ClienteEquipamentoRequestDTO.java
@@ -12,7 +12,7 @@ public record ClienteEquipamentoRequestDTO(
         Integer quantidade,
 
         @NotNull
-        BigDecimal horasUsoDiario,
+        Double horasUsoDiario,
 
         @NotNull
         Integer diasUsoMes
